fix(startup): isolate migration helpers
